FF-106-fix-scrollX-ProfessionModalDesktop: fix scrolling in two desktop components
